INT-2674 - Code Review Changes
