+++ spring-integration-jdbc/src/main/java/org/springframework/integration/jdbc/store/channel/AbstractQueryProvider.java
@@ -10,9 +10,8 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.springframework.integration.jdbc.support.store.channel.query;
+package org.springframework.integration.jdbc.store.channel;
 
-import org.springframework.integration.jdbc.support.store.channel.QueryProvider;
 
 /**
  * @author Gunnar Hillert
