+++ spring-integration-jdbc/src/main/java/org/springframework/integration/jdbc/store/channel/MessageMapper.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.springframework.integration.jdbc.support.store.channel;
+package org.springframework.integration.jdbc.store.channel;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
