+++ spring-integration-jdbc/src/main/java/org/springframework/integration/jdbc/store/channel/PostgresQueryProvider.java
@@ -10,28 +10,26 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.springframework.integration.jdbc.support.store.channel.query;
+package org.springframework.integration.jdbc.store.channel;
 
 /**
  * @author Gunnar Hillert
  * @since 2.2
- *
- * https://blogs.oracle.com/kah/entry/derby_10_5_preview_fetch
  */
-public class DerbyQueryProvider extends AbstractQueryProvider {
+public class PostgresQueryProvider extends AbstractQueryProvider {
 
 	@Override
 	public String getPollFromGroupExcludeIdsQuery() {
 		return "SELECT %PREFIX%CHANNEL_MESSAGE.MESSAGE_ID, %PREFIX%CHANNEL_MESSAGE.MESSAGE_BYTES from %PREFIX%CHANNEL_MESSAGE " +
 				"where %PREFIX%CHANNEL_MESSAGE.GROUP_KEY = :group_key and %PREFIX%CHANNEL_MESSAGE.REGION = :region " +
-				"and %PREFIX%CHANNEL_MESSAGE.MESSAGE_ID not in (:message_ids) order by CREATED_DATE ASC FETCH FIRST ROW ONLY";
+				"and %PREFIX%CHANNEL_MESSAGE.MESSAGE_ID not in (:message_ids) order by CREATED_DATE ASC LIMIT 1 FOR UPDATE";
 	}
 
 	@Override
 	public String getPollFromGroupQuery() {
 		return "SELECT %PREFIX%CHANNEL_MESSAGE.MESSAGE_ID, %PREFIX%CHANNEL_MESSAGE.MESSAGE_BYTES from %PREFIX%CHANNEL_MESSAGE " +
 				"where %PREFIX%CHANNEL_MESSAGE.GROUP_KEY = :group_key and %PREFIX%CHANNEL_MESSAGE.REGION = :region " +
-				"order by CREATED_DATE ASC FETCH FIRST ROW ONLY";
+				"order by CREATED_DATE ASC LIMIT 1 FOR UPDATE";
 	}
 
 }
