+++ spring-integration-jdbc/src/main/java/org/springframework/integration/jdbc/store/channel/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides support classes for the JdbcChannelMessageStore.
+ */
+package org.springframework.integration.jdbc.store.channel;
