+++ spring-integration-jdbc/src/main/java/org/springframework/integration/jdbc/store/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides JDBC-backed Message Store implementations.
+ */
+package org.springframework.integration.jdbc.store;
