+++ spring-integration-jdbc/src/test/java/org/springframework/integration/jdbc/store/channel/DataSource-derby-context.xml
@@ -29,6 +29,6 @@
 	</jdbc:initialize-database>
 
 	<bean id="queryProvider"
-		class="org.springframework.integration.jdbc.support.store.channel.query.DerbyQueryProvider" />
+		class="org.springframework.integration.jdbc.store.channel.DerbyQueryProvider" />
 
 </beans>
