+++ spring-integration-jdbc/src/test/java/org/springframework/integration/jdbc/store/channel/DataSource-postgres-context.xml
@@ -18,6 +18,6 @@
 		<property name="password" value="postgres" />
 	</bean>
 
-	<bean id="queryProvider" class="org.springframework.integration.jdbc.support.store.channel.query.PostgresQueryProvider"/>
+	<bean id="queryProvider" class="org.springframework.integration.jdbc.store.channel.PostgresQueryProvider"/>
 
 </beans>
